refactor(Controller): separate Cart and Product controllers
